==> notebook/5_POIs_extraction.ipynb <==
## POIs extraction 
We extract meaningful POIs from openstreetmap for further scenario analysis by employing the osmnx library

import pandas as pd
import yaml
import osmnx as ox

with open("./conf.yaml") as f:
    conf = yaml.load(f, Loader = yaml.FullLoader)

base_path = conf["base_path"]
data_path = conf["data_path"]

### Beijing subway stations for traffic fluxes monitoring
We get subway stations in a 50km range from Beijing's city center

tags = {"station":"subway"}

gdf = ox.pois.pois_from_point((39.906217,116.3912757), tags, dist = 50000)
#print(gdf.head())
gdf_subway = gdf[(gdf.station == "subway") & (gdf.geometry.notna())]
print("We have {:d} stations in our range".format(len(gdf_subway)))

#we build a dataframe with useful data only (coordinates)
df = pd.DataFrame()
df["name"], df["lat"], df["lon"] = (gdf_subway.name.values, gdf_subway.geometry.centroid.y.values, gdf_subway.geometry.centroid.x.values)
print(df.head())

df.to_csv(data_path + "beijing_subway_stations.csv", index=False)

### Squares, monuments et al.
We extract different kinds of pois in order to model an urban poi-based data collection campaign. 
(We use the same 50km radius)

tags = {"place":"square", "historic":"monument", "shop":"mall", "railway":"station", "amenity":"bus_station"}

gdf = ox.pois.pois_from_point((39.906217,116.3912757), tags, dist = 50000)
#print(gdf.head())
print("We have {:d} POIs in our range".format(len(gdf)))

#print(gdf)
df = pd.DataFrame()
df["name"], df["name:en"], df["lat"], df["lon"] = (gdf.name.values, gdf["name:en"].values, gdf.geometry.centroid.y.values, gdf.geometry.centroid.x.values)
print(df.head())

df.to_csv(data_path+"POIs.csv", index = False)
