==> notebook/6_different_coverage_scenarios.ipynb <==
## Different coverage scenarios
In this notebook we analyze various scenarios regarding different kinds of sensing i.e. environmental monitoring, traffic fluxes and  POI-based sensing.

#we set up all the functions we need and load up some datasets.

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import yaml

with open("conf.yaml") as f:
    conf = yaml.load(f, Loader = yaml.FullLoader)

base_path = conf["base_path"]
data_path = conf["data_path"]
out_path = conf["out_path"]

cols = ["lat", "lon", "uid", "tid","date_time"]

# Define a basic Haversine distance formula
def haversine(lat1, lon1, lat2, lon2):
    
    lat1, lon1, lat2, lon2 = map(np.deg2rad, [lat1, lon1, lat2, lon2])
    dlat = lat2 - lat1 
    dlon = lon2 - lon1 
    a = np.sin(dlat/2)**2 + np.cos(lat1) * np.cos(lat2) * np.sin(dlon/2)**2
    c = 2 * np.arcsin(np.sqrt(a)) 
    #total_meters = METERS * c
    r = 6371000 #radiu * 1000 to return meters
    return c * r

import scipy.integrate as integrate

def coverage(loc_array, tot_dists_df, lam, delta):
    coverage = pd.DataFrame()
    coverage["location"] = loc_array
    coverage["probability"] = calculate_coverage(loc_array, tot_dists_df, lam, delta)
    return coverage

def calculate_coverage(locations, tot_dists_df, lam, delta):
    
    lenght = len(locations)
    count = 0
    coverages = []
    
    #initialize coverage inner function
    lam = lam
    delta = delta
    inner = lambda x: lam * (1/( (np.e)**(lam*x) ))
    integral = lambda x: 1 - integrate.quad(inner, x, x+delta)[0]
    vec_integral = np.vectorize(integral)
    print("Calculating coverage with lambda set to: {:4f} and delta set to: {:f}".format(lam, delta))
    
    for loc in locations:
        
        #if the location has no points whatsoever, assign 0 to the probability and continue
        if(tot_dists_df[tot_dists_df.location == loc].uid.isnull().values.any()):
            coverages.append(0)
            count += 1
            print("Computed coverage on {:d} out of {:d} locations, with lambda set to {:f}".format(count, lenght, lam), end="\r")
            continue
        
        tot_dists_location = tot_dists_df[tot_dists_df.location == loc]
        uids = tot_dists_location.uid.unique()
        #print(uids)
        user_probs = []

        for user in uids:
            #print(user)
            tot_dists_user = tot_dists_location[(tot_dists_location.uid == user)]
            
            #print(tot_dists_user)
            inner_vals = vec_integral(tot_dists_user.distance.values)
            prob = 1 - np.prod(inner_vals)
            user_probs.append(prob)

        location_coverage = 1 - np.prod(user_probs)
        coverages.append(location_coverage)
        count += 1
        print("Computed coverage on {:d} out of {:d} locations, with lambda set to {:f}".format(count, lenght, lam), end="\r")
        
    return coverages

#calculates coverages on multiple lambda values and serializes them on disk
def coverage_multiple_lambdas(lambdas, delta, locations, tot_dists_df, prefix):
    for value in lambdas:
        lam = value
        coverages = coverage(locations, tot_dists_df, lam, delta)
        #merge with df_t dataframe with location positions on "location" id then serialize on disk
        merged = df.merge(coverages, on="location")
        name = prefix + "_coverages_" + str(value).replace(".", ",") + "_lambda.csv"
        merged.to_csv(data_path + name)
        

dataset = pd.read_csv(data_path + "augmented_dataset.csv", usecols=cols, parse_dates = True)
print("Successfully loaded trajectory dataset")



We load up lambda values

#used to parse the conf lambdas
def convert(s):
    try:
        return float(s)
    except ValueError:
        num, denom = s.split('/')
        return float(num) / float(denom)

lambdas = (list(map(convert, conf["lambdas"])))

print(lambdas)

## Traffic fluxes sensing
We load up a csv comprising all of Beijing's subway stations subsequently computing coverage on it after filtering our locations according to a certain detour radius.

#We load up the dataset
df = pd.read_csv(data_path + "beijing_subway_stations.csv")
print(df.info())

#restricting according to detour radius
#for each user, for each trajectory of the user, get the distance of all its waypoints that are in our detour radius' range

detour_radius = conf["detour_radius"] #meters
dists = []
count = 0

for location in df.index.values:
    d_i_h = pd.DataFrame()
    d_i_h["distance"] = haversine(df.loc[location].lat, df.loc[location].lon, dataset["lat"].values, dataset["lon"].values)
    d_i_h = d_i_h[d_i_h.distance < detour_radius]
    if (len(d_i_h) > 0):
        d_i_h["uid"] = dataset["uid"]
        d_i_h["location"] = location
    else:
        d_i_h = pd.DataFrame({"location":location, "uid":np.nan, "distance":np.nan}, index=[0])
        count += 1
        #print(d_i_h)
    dists.append(d_i_h)
    
print("There are {:d} locations that have no points in detour radius' range".format(count)) 

subway_dists = pd.concat(dists)
print(len(subway_dists.location.unique()), " - Unique sensing locations")

#print(tot_dists.location.unique())
locations = subway_dists.location.unique()
df["location"] = locations

print(df)

### Traffic fluxes - coverage probability
Now we compute coverage on the locations according to our model serialize resulting csvs.

#print(lambdas)

delta = conf["delta"]
prefix = "subway" #used to get a meeaningful name for out .csv

print("We're about to calculate coverage with the following lambdas: ", lambdas)

coverage_multiple_lambdas(lambdas, delta, locations, subway_dists, prefix)

## POI-based sensing
We load up a csv with landmarks and similar POIs in order to compute coverage on it, for a sensing based on a wider range of locations

#We load up the dataset
df = pd.read_csv(data_path + "POIs.csv")
print(df.info())

#restricting according to detour radius
#for each user, for each trajectory of the user, get the distance of all its waypoints that are in our detour radius' range

detour_radius = conf["detour_radius"] #meters
dists = []
count = 0

for location in df.index.values:
    d_i_h = pd.DataFrame()
    d_i_h["distance"] = haversine(df.loc[location].lat, df.loc[location].lon, dataset["lat"].values, dataset["lon"].values)
    d_i_h = d_i_h[d_i_h.distance < detour_radius]
    if (len(d_i_h) > 0):
        d_i_h["uid"] = dataset["uid"]
        d_i_h["location"] = location
    else:
        d_i_h = pd.DataFrame({"location":location, "uid":np.nan, "distance":np.nan}, index=[0])
        count += 1
        #print(d_i_h)
    dists.append(d_i_h)
    
print("There are {:d} locations that have no points in detour radius' range".format(count)) 

POIs_dists = pd.concat(dists)
print(len(POIs_dists.location.unique()), " - Unique sensing locations")

#print(tot_dists.location.unique())
locations = POIs_dists.location.unique()
df["location"] = locations

print(df)

## POIs - coverage probability

#print(lambdas)

delta = conf["delta"]
prefix = "POIs" #used to get a meeaningful name for out .csv

print("We're about to calculate coverage with the following lambdas: ", lambdas)

coverage_multiple_lambdas(lambdas, delta, locations, POIs_dists, prefix)

## Grid-based environmental sensing
We load up a previously generated regular grid in order to simulate an environmental type of sensing, (e.g. environmental values such as temperature, pressure etc.)

#We load up the grid dataset
df = pd.read_csv(data_path + "5000_m_grid.csv")
print(df.info())

#restricting according to detour radius
#for each user, for each trajectory of the user, get the distance of all its waypoints that are in our detour radius' range

detour_radius = conf["detour_radius"] #meters
dists = []
count = 0

for location in df.index.values:
    d_i_h = pd.DataFrame()
    d_i_h["distance"] = haversine(df.loc[location].lat, df.loc[location].lon, dataset["lat"].values, dataset["lon"].values)
    d_i_h = d_i_h[d_i_h.distance < detour_radius]
    if (len(d_i_h) > 0):
        d_i_h["uid"] = dataset["uid"]
        d_i_h["location"] = location
    else:
        d_i_h = pd.DataFrame({"location":location, "uid":np.nan, "distance":np.nan}, index=[0])
        count += 1
        #print(d_i_h)
    dists.append(d_i_h)
    
print("There are {:d} locations that have no points in detour radius' range".format(count)) 

grid_dists = pd.concat(dists)
print(len(grid_dists.location.unique()), " - Unique sensing locations")

#print(tot_dists.location.unique())
locations = grid_dists.location.unique()
df["location"] = locations

print(df)

### Grid - coverage probability

#print(lambdas)

delta = conf["delta"]
prefix = "grid" #used to get a meeaningful name for out .csv

print("We're about to calculate coverage with the following lambdas: ", lambdas)

coverage_multiple_lambdas(lambdas, delta, locations, grid_dists, prefix)
